+++ EDDiscovery/EliteDangerous/JournalEvents/JournalMarketSell.cs
@@ -20,18 +20,19 @@ namespace EDDiscovery.EliteDangerous.JournalEvents
         {
             Type = Tools.GetStringDef(evt["Type"]);
             Count = Tools.GetInt(evt["Count"]);
-            SellPrice = Tools.GetInt(evt["SellPrice"]);
-            TotalSale = Tools.GetInt(evt["TotalSale"]);
-            AvgPricePaid = Tools.GetInt(evt["AvgPricePaid"]);
+            SellPrice = Tools.GetLong(evt["SellPrice"]);
+            TotalSale = Tools.GetLong(evt["TotalSale"]);
+            AvgPricePaid = Tools.GetLong(evt["AvgPricePaid"]);
             IllegalGoods = Tools.GetBool(evt["IllegalGoods"]);
             StolenGoods = Tools.GetBool(evt["StolenGoods"]);
             BlackMarket = Tools.GetBool(evt["BlackMarket"]);
         }
+
         public string Type { get; set; }
         public int Count { get; set; }
-        public int SellPrice { get; set; }
-        public int TotalSale { get; set; }
-        public int AvgPricePaid { get; set; }
+        public long SellPrice { get; set; }
+        public long TotalSale { get; set; }
+        public long AvgPricePaid { get; set; }
         public bool IllegalGoods { get; set; }
         public bool StolenGoods { get; set; }
         public bool BlackMarket { get; set; }
